dev: split to bundle_stats and u256_numeric_string
