+++ crates/rpc-types-mev/src/lib.rs
@@ -6,6 +6,10 @@
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
+// serde helper to serialize/deserialize u256 as numeric string
+mod bundle_stats;
+mod u256_numeric_string;
+
 use alloy_eips::{BlockId, BlockNumberOrTag};
 use alloy_primitives::{Address, Bytes, Log, TxHash, B256, U256};
 use serde::{
@@ -44,11 +48,7 @@ pub struct Inclusion {
     #[serde(with = "alloy_serde::quantity")]
     pub block: u64,
     /// The last block the bundle is valid for.
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub max_block: Option<u64>,
 }
 
@@ -335,32 +335,16 @@ pub struct SimBundleOverrides {
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub coinbase: Option<Address>,
     /// Timestamp used for simulation, defaults to parentBlock.timestamp + 12
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub timestamp: Option<u64>,
     /// Gas limit used for simulation, defaults to parentBlock.gasLimit
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub gas_limit: Option<u64>,
     /// Base fee used for simulation, defaults to parentBlock.baseFeePerGas
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub base_fee: Option<u64>,
     /// Timeout in seconds, defaults to 5
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub timeout: Option<u64>,
 }
 
@@ -439,11 +423,7 @@ pub struct PrivateTransactionRequest {
     pub tx: Bytes,
     /// Hex-encoded number string, optional. Highest block number in which the transaction should
     /// be included.
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub max_block_number: Option<u64>,
     /// Preferences for private transaction.
     #[serde(default, skip_serializing_if = "PrivateTransactionPreferences::is_empty")]
@@ -474,99 +454,6 @@ impl PrivateTransactionPreferences {
 pub struct CancelPrivateTransactionRequest {
     /// Transaction hash of the transaction to be canceled
     pub tx_hash: B256,
-}
-
-// TODO(@optimiz-r): Revisit after <https://github.com/flashbots/flashbots-docs/issues/424> is closed.
-/// Response for `flashbots_getBundleStatsV2` represents stats for a single bundle
-///
-/// Note: this is V2: <https://docs.flashbots.net/flashbots-auction/searchers/advanced/rpc-endpoint#flashbots_getbundlestatsv2>
-///
-/// Timestamp format: "2022-10-06T21:36:06.322Z"
-#[derive(Default, Debug, Clone, PartialEq, Eq)]
-pub enum BundleStats {
-    /// The relayer has not yet seen the bundle.
-    #[default]
-    Unknown,
-    /// The relayer has seen the bundle, but has not simulated it yet.
-    Seen(StatsSeen),
-    /// The relayer has seen the bundle and has simulated it.
-    Simulated(StatsSimulated),
-}
-
-impl Serialize for BundleStats {
-    fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        match self {
-            Self::Unknown => serde_json::json!({"isSimulated": false}).serialize(serializer),
-            Self::Seen(stats) => stats.serialize(serializer),
-            Self::Simulated(stats) => stats.serialize(serializer),
-        }
-    }
-}
-
-impl<'de> Deserialize<'de> for BundleStats {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let map = serde_json::Map::deserialize(deserializer)?;
-
-        if map.get("receivedAt").is_none() {
-            Ok(Self::Unknown)
-        } else if map["isSimulated"] == false {
-            StatsSeen::deserialize(serde_json::Value::Object(map))
-                .map(BundleStats::Seen)
-                .map_err(serde::de::Error::custom)
-        } else {
-            StatsSimulated::deserialize(serde_json::Value::Object(map))
-                .map(BundleStats::Simulated)
-                .map_err(serde::de::Error::custom)
-        }
-    }
-}
-
-/// Response for `flashbots_getBundleStatsV2` represents stats for a single bundle
-///
-/// Note: this is V2: <https://docs.flashbots.net/flashbots-auction/searchers/advanced/rpc-endpoint#flashbots_getbundlestatsv2>
-///
-/// Timestamp format: "2022-10-06T21:36:06.322Z
-#[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct StatsSeen {
-    /// boolean representing if this searcher has a high enough reputation to be in the high
-    /// priority queue
-    pub is_high_priority: bool,
-    /// representing whether the bundle gets simulated. All other fields will be omitted except
-    /// simulated field if API didn't receive bundle
-    pub is_simulated: bool,
-    /// time at which the bundle API received the bundle
-    pub received_at: String,
-}
-
-/// Response for `flashbots_getBundleStatsV2` represents stats for a single bundle
-///
-/// Note: this is V2: <https://docs.flashbots.net/flashbots-auction/searchers/advanced/rpc-endpoint#flashbots_getbundlestatsv2>
-///
-/// Timestamp format: "2022-10-06T21:36:06.322Z
-#[derive(Default, Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct StatsSimulated {
-    /// boolean representing if this searcher has a high enough reputation to be in the high
-    /// priority queue
-    pub is_high_priority: bool,
-    /// representing whether the bundle gets simulated. All other fields will be omitted except
-    /// simulated field if API didn't receive bundle
-    pub is_simulated: bool,
-    /// time at which the bundle gets simulated
-    pub simulated_at: String,
-    /// time at which the bundle API received the bundle
-    pub received_at: String,
-    /// indicates time at which each builder selected the bundle to be included in the target
-    /// block
-    #[serde(default = "Vec::new")]
-    pub considered_by_builders_at: Vec<ConsideredByBuildersAt>,
-    /// indicates time at which each builder sealed a block containing the bundle
-    #[serde(default = "Vec::new")]
-    pub sealed_by_builders_at: Vec<SealedByBuildersAt>,
 }
 
 /// Represents information about when a bundle was considered by a builder.
@@ -635,18 +522,10 @@ pub struct EthSendBundle {
     #[serde(with = "alloy_serde::quantity")]
     pub block_number: u64,
     /// unix timestamp when this bundle becomes active
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub min_timestamp: Option<u64>,
     /// unix timestamp how long this bundle stays valid
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub max_timestamp: Option<u64>,
     /// list of hashes of possibly reverting txs
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
@@ -678,11 +557,7 @@ pub struct EthCallBundle {
     /// Either a hex encoded number or a block tag for which state to base this simulation on
     pub state_block_number: BlockNumberOrTag,
     /// the timestamp to use for this bundle simulation, in seconds since the unix epoch
-    #[serde(
-        default,
-        with = "alloy_serde::quantity::opt",
-        skip_serializing_if = "Option::is_none"
-    )]
+    #[serde(default, with = "alloy_serde::quantity::opt", skip_serializing_if = "Option::is_none")]
     pub timestamp: Option<u64>,
 }
 
@@ -749,42 +624,10 @@ pub struct EthCallBundleTransactionResult {
     pub revert: Option<Bytes>,
 }
 
-mod u256_numeric_string {
-    use alloy_primitives::U256;
-    use serde::{de, Deserialize, Serializer};
-    use std::str::FromStr;
-
-    pub(crate) fn deserialize<'de, D>(deserializer: D) -> Result<U256, D::Error>
-    where
-        D: de::Deserializer<'de>,
-    {
-        let val = serde_json::Value::deserialize(deserializer)?;
-        match val {
-            serde_json::Value::String(s) => {
-                if let Ok(val) = s.parse::<u128>() {
-                    return Ok(U256::from(val));
-                }
-                U256::from_str(&s).map_err(de::Error::custom)
-            }
-            serde_json::Value::Number(num) => {
-                num.as_u64().map(U256::from).ok_or_else(|| de::Error::custom("invalid u256"))
-            }
-            _ => Err(de::Error::custom("invalid u256")),
-        }
-    }
-
-    pub(crate) fn serialize<S>(val: &U256, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let val: u128 = (*val).try_into().map_err(serde::ser::Error::custom)?;
-        serializer.serialize_str(&val.to_string())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::bundle_stats::BundleStats;
     use std::str::FromStr;
 
     #[test]
@@ -977,78 +820,5 @@ mod tests {
   }"#;
 
         let _call = serde_json::from_str::<EthCallBundleResponse>(s).unwrap();
-    }
-
-    #[test]
-    fn can_serialize_deserialize_bundle_stats() {
-        let fixtures = [
-            (
-                r#"{
-                    "isSimulated": false
-                }"#,
-                BundleStats::Unknown,
-            ),
-            (
-                r#"{
-                    "isHighPriority": false,
-                    "isSimulated": false,
-                    "receivedAt": "476190476193"
-                }"#,
-                BundleStats::Seen(StatsSeen {
-                    is_high_priority: false,
-                    is_simulated: false,
-                    received_at: "476190476193".to_string(),
-                }),
-            ),
-            (
-                r#"{
-                    "isHighPriority": true,
-                    "isSimulated": true,
-                    "simulatedAt": "111",
-                    "receivedAt": "222",
-                    "consideredByBuildersAt":[],
-                    "sealedByBuildersAt": [
-                        {
-                            "pubkey": "333",
-                            "timestamp": "444"
-                        },
-                        {
-                            "pubkey": "555",
-                            "timestamp": "666"
-                        }
-                    ]
-                }"#,
-                BundleStats::Simulated(StatsSimulated {
-                    is_high_priority: true,
-                    is_simulated: true,
-                    simulated_at: String::from("111"),
-                    received_at: String::from("222"),
-                    considered_by_builders_at: vec![],
-                    sealed_by_builders_at: vec![
-                        SealedByBuildersAt {
-                            pubkey: String::from("333"),
-                            timestamp: String::from("444"),
-                        },
-                        SealedByBuildersAt {
-                            pubkey: String::from("555"),
-                            timestamp: String::from("666"),
-                        },
-                    ],
-                }),
-            ),
-        ];
-
-        let strip_whitespaces =
-            |input: &str| input.chars().filter(|&c| !c.is_whitespace()).collect::<String>();
-
-        for (serialized, deserialized) in fixtures {
-            // Check de-serialization
-            let deserialized_expected = serde_json::from_str::<BundleStats>(serialized).unwrap();
-            assert_eq!(deserialized, deserialized_expected);
-
-            // Check serialization
-            let serialized_expected = &serde_json::to_string(&deserialized).unwrap();
-            assert_eq!(strip_whitespaces(serialized), strip_whitespaces(serialized_expected));
-        }
     }
 }
